Fix hero bubbles and hero element sizing on mobile
When on mobile the height of the hero element and hero bubble effect
would load with different height in chrome if you had multiple tabs in
a group.

Due to css 100vh and js window.height being calculated differently.
Window.height taking into account the tab bar and the 100vh not.