Fix mobile scroll position when form submission
After the contact form is submitted the scroll position should scroll
to the top of the form instead it reloaded the page at very top of the
site or at the point where the user submitted the form.

Due to the function controlling the scroll position triggering
quicker than the target scroll position could be set. When the scroll
position was returned to the very top of the site it was due to the
anchor point extension being different in the url of the rendered page
after the form is submitted.